requirements.txt
blinkt==0.1.2
Flask==3.0.1
Flask_APScheduler==1.13.1
numpy==1.24.2
opencv_python==4.9.0.80
picamera2==0.3.16
Pillow==10.2.0
pytesseract==0.3.10
PyYAML==6.0.1
RPi.GPIO==0.7.1a4
Wand==0.6.13
gunicorn==21.2.0
